Sequential parameter estimation for the calibration of complex models

This function performs the optimization of a function, possibly in sequential phases of increasing complexity, and it is designed for the calibration of a model, by minimizing the error function fn associated to it.

calibrate(
  par,
  fn,
  gr,
  ...,
  lower,
  upper,
  phases,
  method,
  control,
  hessian,
  replicates,
  parallel
)

Arguments

par

A numeric vector or list. The length of the par argument defines the number of parameters to be estimated (i.e. the dimension of the problem).

fn

The function to be minimized.

gr

the gradient of fn. Ignored, added for portability with other optimization functions.

...

Additional parameters to be passed to fn.

lower

Lower threshold value(s) for parameters. One value or a vector of the same length as par. If one value is provided, it is used for all parameters. NA means -Inf. By default -Inf is used (unconstrained).

upper

Upper threshold value(s) for parameters. One value or a vector of the same length as par. If one value is provided, it is used for all parameters. NA means Inf. By default Inf is used (unconstrained).

phases

An optional vector of the same length as par, indicating the phase at which each parameter becomes active. If omitted, default value is 1 for all parameters, performing a single optimization.

method

The optimization method to be used. The default method is the AHR-ES (Adaptative Hierarchical Recombination Evolutionary Strategy, Oliveros-Ramos & Shin, 2016). See details for the methods available.

control

Parameter for the control of the algorithm itself, see details.

hessian

Logical. Should a numerically differentiated Hessian matrix be returned? Currently not implemented.

replicates

The number of replicates for the evaluation of fn. The default value is 1. A value greater than 1 is only useful for stochastic functions.

parallel

Logical. Use parallel computation of gradients?

Details

In the control list, aggFn is a function to aggregate fn to a scalar value if the returned value is a vector. Some optimization algorithm can exploite the additional information provided by a vectorial output from fn.

See also

Author

Ricardo Oliveros-Ramos

Examples

calibrate(par=rep(NA, 5), fn=sphereN)
#> 
#> Phase 1 finished in 0.62s (5 of 5 parameters active)
#> Function value: 0.223404
#> Parameter values: 0.0701 -0.0312 -0.168 0.0158 0.0158
#> 
#> Calibration finished.
#> Function value: 0.2234045 
#> Parameters:
#> [1]  0.07014151 -0.03120345 -0.16843871  0.01580654  0.01582654
if (FALSE) {
calibrate(par=rep(NA, 5), fn=sphereN, replicates=3)
calibrate(par=rep(0.5, 5), fn=sphereN, replicates=3, lower=-5, upper=5)
calibrate(par=rep(0.5, 5), fn=sphereN, replicates=3, lower=-5, upper=5, phases=c(1,1,1,2,3))
calibrate(par=rep(0.5, 5), fn=sphereN, replicates=c(1,1,4), lower=-5, upper=5, phases=c(1,1,1,2,3))
}